Improved display of generated PDF
Set the title and author fields of PDF and fixed PDF bookmarks to
display all the heading levels I use (by including them in the global
table of contents). Also set links to not be surrounded by distracting
boxes but to be called out by being different colors instead. I may
want to make this color dark blue so it doesn't interfere with
printing.